what i have done so far on the title screen!
have some simple button and menu stuff implemented, but hoping to make the buttons more interactable (like highlighted on hover), but I'm not sure how to recolor sprites... Also don't know scene switching for when you press play, so someone else may need to help with that. The credits screen is a placeholder for now, I'll make a new one when everyone gives their info. I'm guessing we won't have an options menu implemented by the demo, so it's just a "coming soon" for now!